requirements.txt
streamlit
openai
python-dotenv
pandas
azure-storage-blob